refactor: MailTemplate 관련 template 패키지 바로 밑으로 이동 #98
